\name{BasicEconometrics-package}
\alias{BasicEconometrics-package}
\alias{BasicEconometrics}
\docType{package}
\title{
\packageTitle{BasicEconometrics}
}
\description{
\packageDescription{BasicEconometrics}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BasicEconometrics}
\packageIndices{BasicEconometrics}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{BasicEconometrics}

Maintainer: \packageMaintainer{BasicEconometrics}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
