% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Times of a atheletes before and after a training camp}
\format{A data frame with 12 rows and 3 columns:
\describe{
\item{AthleteID}{Athlete identification, by numbers}
\item{Time1}{Time before the training, in minutes}
\item{Time2}{Time after the training, in minutes}
}}
\usage{
data
}
\description{
a data set containing the times before and after of a group
of athletes before and after receiving a training
}
\keyword{datasets}
