% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Durbin_Watson_test_and_Cochrane_Orcutt_Correction.R
\name{durbin_watson_test_and_Cochr_orcutt}
\alias{durbin_watson_test_and_Cochr_orcutt}
\title{durbin_watson_test_and_Cochr_orcutt}
\usage{
durbin_watson_test_and_Cochr_orcutt(reg_list, alpha = 0.05)
}
\description{
realiza o teste de durbin watson de autocorrela??o residual e caso ela (autocorr) seja
significativa ent?o estima-se novamente a regress?o atrav?s corrigindo a autocorrela??o por
meio do m?todo iterativo de cochrane orcutt
}
